refactor(oracle): split cli subcommand parsers into command modules
